requirements.txt
Pillow==8.4.0
torch==1.10.2
torchvision==0.11.3
